requirements.txt
Pillow==11.2.1
tqdm==4.67.0
